fixing warnings, removing unused code, refs #3, #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@14332 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/PositionVector.cpp
@@ -1040,7 +1040,7 @@ PositionVector::insertAt(int index, const Position& p) {
 void
 PositionVector::replaceAt(int index, const Position& p) {
     assert(index < static_cast<int>(size()));
-    assert(index + size() >= 0);
+    assert(index + static_cast<int>(size()) >= 0);
     if (index >= 0) {
         (*this)[index] = p;
     } else {
